% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_mapping.R
\name{convert_links}
\alias{convert_links}
\title{Convert links}
\usage{
convert_links(x)
}
\arguments{
\item{x}{a data.frame or tibble}
}
\value{
the same data frame with any http text converted to hyperlink html.
}
\description{
Given a data.frame/ tibble, converts any
texts in a data frame/ tibble that begin with "httr" to
a hyperlink
}
