% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_GET_methods_Locations.R
\name{sta_getLocations}
\alias{sta_getLocations}
\title{Retrieve all locations from STA instance}
\usage{
sta_getLocations(api)
}
\arguments{
\item{api}{character URL of an STA endpoint.}
}
\value{
an sf object of every STA location entity.
}
\description{
Given an STA endpoint, will get all location entities and
collate them into an \href{https://cran.r-project.org/web/packages/sf/}{\code{sf}} object.
It checks if the STA call has been paginated. If it is paginated, it retrieves
the total number of locations and the default $skip value, and iterates through,
retrieving all locations.
}
