% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_GET_methods_Locations.R
\name{sta_getLocationsCall}
\alias{sta_getLocationsCall}
\title{Parse a GET Locations request response from an STA instance}
\usage{
sta_getLocationsCall(call)
}
\arguments{
\item{call}{character URL of an STA Locations call, including any query parameters}
}
\value{
an sf object of every STA location entity.
}
\description{
Given an STA/Locations call, including any query parameters,
will get all location entities collate into an
\href{https://cran.r-project.org/web/packages/sf/}{\code{sf}} object.
A helper function for \code{sta_getLocations} as it iterates through a paginated
STA Location document.
}
