% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{sta_iterate}
\alias{sta_iterate}
\title{Iterator}
\usage{
sta_iterate(call, response, fun)
}
\arguments{
\item{call}{character URL of STA call.}

\item{fun}{any function that parses STA responses
into R data structures that can be row-binded.
An example is \code{sta_getLocationsCall()}}

\item{resonse}{json document corresponding to the
\code{call}}
}
\value{
a collation of objects created by \code{fun}
}
\description{
Given an STA call, response document,
and parsing function, will detect the total entity count
and default $skip value. It will use these values to
iterate through the paginated response document, and
collate together the resulting responses according
to the passed function
}
