% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta_mapping.R
\name{sta_mapview}
\alias{sta_mapview}
\title{Map locations of SensorThings API Locations entities}
\usage{
sta_mapview(api, ...)
}
\arguments{
\item{api}{a character URL for an STA endpoint}

\item{...}{any arguments for \code{mapview}}
}
\value{
\code{mapview} plot
}
\description{
Given a STA endpoint, creates a \code{mapview} html map
of the endpoints and any links provided by STA.
}
